--- calculators/src/chemistry/chemicalKineticsCalc/MainChemicalKineticsCalculator.js
@@ -1,95 +1,0 @@
-import React, { Component } from 'react';
-import { Container, Typography, TextField, Button, Box } from '@mui/material';
-
-class ChemicalKineticsCalculator extends Component {
-  constructor() {
-    super();
-    this.state = {
-      reactantAInitialConcentration: '',
-      reactantBInitialConcentration: '',
-      productInitialConcentration: '',
-      time: '',
-      reactionRate: null,
-      rateConstant: null,
-    };
-  }
-
-  // Handle input changes in the form fields
-  handleInputChange = (event) => {
-    const { name, value } = event.target;
-    this.setState({ [name]: value });
-  }
-
-  // Function to calculate the reaction rate and rate constant
-  calculateRate = () => {
-    // Implement your chemical kinetics calculations here.
-    const reactionRate = Math.random() * 10;
-    const rateConstant = Math.random() * 5;
-    this.setState({ reactionRate, rateConstant });
-  }
-
-  render() {
-    return (
-      <Container maxWidth="sm">
-        {/* Application title */}
-        <Typography variant="h4" gutterBottom>
-          Chemical Kinetics Calculator
-        </Typography>
-
-        {/* Input fields for initial concentrations and time */}
-        <TextField
-          label="Initial Concentration of Reactant A"
-          type="number"
-          name="reactantAInitialConcentration"
-          value={this.state.reactantAInitialConcentration}
-          onChange={this.handleInputChange}
-          fullWidth
-          margin="normal"
-        />
-        <TextField
-          label="Initial Concentration of Reactant B"
-          type="number"
-          name="reactantBInitialConcentration"
-          value={this.state.reactantBInitialConcentration}
-          onChange={this.handleInputChange}
-          fullWidth
-          margin="normal"
-        />
-        <TextField
-          label="Initial Concentration of Product"
-          type="number"
-          name="productInitialConcentration"
-          value={this.state.productInitialConcentration}
-          onChange={this.handleInputChange}
-          fullWidth
-          margin="normal"
-        />
-        <TextField
-          label="Time"
-          type="number"
-          name="time"
-          value={this.state.time}
-          onChange={this.handleInputChange}
-          fullWidth
-          margin="normal"
-        />
-
-        {/* Calculate button */}
-        <Button variant="contained" onClick={this.calculateRate} color="primary" fullWidth>
-          Calculate Rate
-        </Button>
-
-        {/* Display results if available */}
-        {this.state.reactionRate !== null && (
-          <Box marginTop={2}>
-            <Typography variant="h6">Results:</Typography>
-            <Typography>Reaction Rate: {this.state.reactionRate}</Typography>
-            <Typography>Rate Constant: {this.state.rateConstant}</Typography>
-          </Box>
-        )}
-      </Container>
-    );
-  }
-}
-
-export default ChemicalKineticsCalculator;
